test: move per-package *.smoke.ts into a smoke/ subdir
The smoke tests were co-located with src at each package root, cluttering the
package dirs. Move every *.smoke.ts into a per-package smoke/ subdirectory
(packages/core/smoke, extensions/connector-core/smoke, implementations/{cli,
manager}/smoke, extensions/connector-{hermes,opencode}/smoke) and fix their
relative imports (./src -> ../src) + the package.json tsx paths. No behavior
change; full sweep + typecheck green from the new locations.
