+++ extensions/connector-core/smoke/per-channel-attention.smoke.ts
@@ -14,9 +14,9 @@ import { mkdtempSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { CotalEndpoint, isReachable, seedChannelRegistry } from "@cotal-ai/core";
-import { MeshAgent } from "./src/agent.js";
-import type { AgentConfig } from "./src/config.js";
-import type { InboxItem } from "./src/agent.js";
+import { MeshAgent } from "../src/agent.js";
+import type { AgentConfig } from "../src/config.js";
+import type { InboxItem } from "../src/agent.js";
 
 const PORT = 14241;
 const servers = `nats://127.0.0.1:${PORT}`;
