+++ package.json
@@ -23,33 +23,33 @@
     "ci:publish": "pnpm build && pnpm publish -r --provenance --access=public --no-git-checks",
     "smoke": "tsx packages/core/smoke.ts",
     "smoke:auth": "tsx packages/core/smoke-auth.ts",
-    "smoke:view": "tsx implementations/cli/view.smoke.ts",
-    "smoke:members": "tsx packages/core/members.smoke.ts",
-    "smoke:plane3:auth": "tsx packages/core/plane3-auth.smoke.ts",
-    "smoke:plane3-gate:auth": "tsx packages/core/plane3-gate-auth.smoke.ts",
-    "smoke:membership": "tsx packages/core/membership.smoke.ts",
-    "smoke:membership:auth": "tsx packages/core/membership-auth.smoke.ts",
-    "smoke:channels": "tsx packages/core/channels.smoke.ts",
-    "smoke:wildcard-backfill": "tsx packages/core/wildcard-backfill.smoke.ts",
-    "smoke:channels:auth": "tsx packages/core/channels-auth.smoke.ts",
-    "smoke:read-acl:auth": "tsx packages/core/read-acl-auth.smoke.ts",
-    "smoke:sub-acl:auth": "tsx packages/core/sub-acl-auth.smoke.ts",
-    "smoke:self-serve-join:auth": "tsx packages/core/self-serve-join-auth.smoke.ts",
-    "smoke:self-serve-join-coverage:auth": "tsx packages/core/self-serve-join-coverage-auth.smoke.ts",
-    "smoke:e2e:acl": "tsx packages/core/e2e-acl.smoke.ts",
-    "smoke:control-auth": "tsx packages/core/control-auth.smoke.ts",
-    "smoke:attention": "tsx extensions/connector-core/attention.smoke.ts",
-    "smoke:attention:auth": "tsx extensions/connector-core/attention-auth.smoke.ts",
-    "smoke:channel-attention": "tsx extensions/connector-core/per-channel-attention.smoke.ts",
-    "smoke:channel-attention:auth": "tsx extensions/connector-core/per-channel-attention-auth.smoke.ts",
-    "smoke:presence-scrub": "tsx packages/core/presence-offline-scrub.smoke.ts",
-    "smoke:cross-path-dedup": "tsx extensions/connector-core/cross-path-dedup.smoke.ts",
-    "smoke:feedback": "tsx extensions/connector-core/feedback.smoke.ts",
-    "smoke:opencode": "tsx extensions/connector-opencode/turn-wedge.smoke.ts",
-    "smoke:attach": "tsx implementations/manager/attach.smoke.ts",
-    "smoke:env-isolate": "tsx implementations/manager/env-isolate.smoke.ts",
-    "smoke:install": "tsx implementations/cli/install.smoke.ts",
-    "smoke:send": "tsx implementations/cli/send.smoke.ts"
+    "smoke:view": "tsx implementations/cli/smoke/view.smoke.ts",
+    "smoke:members": "tsx packages/core/smoke/members.smoke.ts",
+    "smoke:plane3:auth": "tsx packages/core/smoke/plane3-auth.smoke.ts",
+    "smoke:plane3-gate:auth": "tsx packages/core/smoke/plane3-gate-auth.smoke.ts",
+    "smoke:membership": "tsx packages/core/smoke/membership.smoke.ts",
+    "smoke:membership:auth": "tsx packages/core/smoke/membership-auth.smoke.ts",
+    "smoke:channels": "tsx packages/core/smoke/channels.smoke.ts",
+    "smoke:wildcard-backfill": "tsx packages/core/smoke/wildcard-backfill.smoke.ts",
+    "smoke:channels:auth": "tsx packages/core/smoke/channels-auth.smoke.ts",
+    "smoke:read-acl:auth": "tsx packages/core/smoke/read-acl-auth.smoke.ts",
+    "smoke:sub-acl:auth": "tsx packages/core/smoke/sub-acl-auth.smoke.ts",
+    "smoke:self-serve-join:auth": "tsx packages/core/smoke/self-serve-join-auth.smoke.ts",
+    "smoke:self-serve-join-coverage:auth": "tsx packages/core/smoke/self-serve-join-coverage-auth.smoke.ts",
+    "smoke:e2e:acl": "tsx packages/core/smoke/e2e-acl.smoke.ts",
+    "smoke:control-auth": "tsx packages/core/smoke/control-auth.smoke.ts",
+    "smoke:attention": "tsx extensions/connector-core/smoke/attention.smoke.ts",
+    "smoke:attention:auth": "tsx extensions/connector-core/smoke/attention-auth.smoke.ts",
+    "smoke:channel-attention": "tsx extensions/connector-core/smoke/per-channel-attention.smoke.ts",
+    "smoke:channel-attention:auth": "tsx extensions/connector-core/smoke/per-channel-attention-auth.smoke.ts",
+    "smoke:presence-scrub": "tsx packages/core/smoke/presence-offline-scrub.smoke.ts",
+    "smoke:cross-path-dedup": "tsx extensions/connector-core/smoke/cross-path-dedup.smoke.ts",
+    "smoke:feedback": "tsx extensions/connector-core/smoke/feedback.smoke.ts",
+    "smoke:opencode": "tsx extensions/connector-opencode/smoke/turn-wedge.smoke.ts",
+    "smoke:attach": "tsx implementations/manager/smoke/attach.smoke.ts",
+    "smoke:env-isolate": "tsx implementations/manager/smoke/env-isolate.smoke.ts",
+    "smoke:install": "tsx implementations/cli/smoke/install.smoke.ts",
+    "smoke:send": "tsx implementations/cli/smoke/send.smoke.ts"
   },
   "dependencies": {
     "@cotal-ai/cli": "workspace:*",
