+++ packages/core/smoke/members.smoke.ts
@@ -28,7 +28,7 @@ import {
   durableEligible,
   StaleMembershipWrite,
   type MembershipRecord,
-} from "./src/index.js";
+} from "../src/index.js";
 
 const PORT = 20000 + Math.floor(Math.random() * 40000);
 const servers = `nats://127.0.0.1:${PORT}`;
