+++ packages/core/smoke/plane3-auth.smoke.ts
@@ -34,7 +34,7 @@ import {
   dlvStream,
   type Delivery,
   type MessageMeta,
-} from "./src/index.js";
+} from "../src/index.js";
 
 const PORT = 20000 + Math.floor(Math.random() * 40000);
 const SERVERS = `nats://127.0.0.1:${PORT}`;
