+++ packages/core/smoke/self-serve-join-coverage-auth.smoke.ts
@@ -39,7 +39,7 @@ import {
   type CotalMessage,
   type Delivery,
   type MessageMeta,
-} from "./src/index.js";
+} from "../src/index.js";
 
 const PORT = 20000 + Math.floor(Math.random() * 40000);
 const SERVERS = `nats://127.0.0.1:${PORT}`;
