+++ packages/core/smoke/sub-acl-auth.smoke.ts
@@ -23,7 +23,7 @@ import {
   setupSpaceStreams,
   chatSubject,
   spacePrefix,
-} from "./src/index.js";
+} from "../src/index.js";
 
 const PORT = 14243;
 const SERVERS = `nats://127.0.0.1:${PORT}`;
